Service/HTTP: Moved declarations to the header file
